requirements.txt
numpy>=1.23,<2.0
matplotlib>=3.6,<4.0
torch>=2.1,<3.0
torchvision>=0.16,<1.0
tqdm>=4.65,<5.0
jupyter>=1.0
torchmetrics[image]>=1.0
scipy>=1.10
pillow>=9.0
